requirements.txt
streamlit
openai
firebase-admin==6.5.0
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0
